Add Continuous Integration via Github Actions
Add a CI workflow for Github Actions. It runs "mvn package" on the project, ensuring that the commit passes all the tests
This is more of an easy review thing, rather than proper CI as we don't have any limits on merging between branches etc
It runs automatically on every commit to any branch. The results can be viewed at https://github.com/Kwandes/AdventureXP/actions

A "Secret" is just a Github term for an Environment Variable. It works the same way our application.properties JDBC URL does